Create SectionRegulation dedicated package
