+++ wildlifecompliance/components/offence/admin.py
@@ -2,8 +2,9 @@ from django import forms
 from django.contrib import admin
 from reversion.admin import VersionAdmin
 
+import wildlifecompliance.components.section_regulation.models
 from wildlifecompliance.components.offence import models
-from wildlifecompliance.components.offence.models import PenaltyAmount
+from wildlifecompliance.components.section_regulation.models import PenaltyAmount
 
 
 class PenaltyAmountInline(admin.TabularInline):
@@ -21,7 +22,7 @@ class SectionRegulationForm(forms.ModelForm):
     offence_text = forms.CharField(widget=forms.Textarea)
 
     class Meta:
-        model = models.SectionRegulation
+        model = wildlifecompliance.components.section_regulation.models.SectionRegulation
         fields = '__all__'
 
 
@@ -30,10 +31,10 @@ class SectionRegulationAdmin(VersionAdmin):
     inlines = [PenaltyAmountInline,]
 
 
-@admin.register(models.PenaltyAmount)
+@admin.register(wildlifecompliance.components.section_regulation.models.PenaltyAmount)
 class PenaltyAmountAdmin(admin.ModelAdmin):
     pass
 
 
-admin.site.register(models.SectionRegulation, SectionRegulationAdmin)
+admin.site.register(wildlifecompliance.components.section_regulation.models.SectionRegulation, SectionRegulationAdmin)
